integration: rename 'matcher utils' file
